+++ data-protocols/dsp/dsp-http-core/build.gradle.kts
@@ -19,11 +19,13 @@ plugins {
 dependencies {
     api(project(":spi:common:http-spi"))
     api(project(":spi:common:json-ld-spi"))
+    api(project(":spi:common:validator-spi"))
     api(project(":spi:control-plane:contract-spi"))
     api(project(":spi:control-plane:transfer-spi"))
     api(project(":extensions:common:json-ld"))
     api(project(":data-protocols:dsp:dsp-spi"))
     api(project(":data-protocols:dsp:dsp-http-spi"))
-
+    
+    testImplementation(project(":extensions:common:http"))
     testImplementation(project(":core:common:junit"))
 }
