+++ data-protocols/dsp/dsp-http-core/src/test/java/org/eclipse/edc/protocol/dsp/message/DspRequestHandlerImplTest.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.protocol.dsp.api.configuration.message;
+package org.eclipse.edc.protocol.dsp.message;
 
 import jakarta.json.Json;
 import jakarta.json.JsonObject;
